+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 
 ### Added
 - `ExecutionMiddleware\ReadOnlyOperationMiddleware` rejects GET requests targeting mutations [\#1261 / mfn](https://github.com/rebing/graphql-laravel/pull/1261)
+- `Support\Middleware\CsrfGuard` CSRF protection middleware [\#1265 / mfn](https://github.com/rebing/graphql-laravel/pull/1265)
 
 ### Changed
 - `ExecutionMiddleware\AddAuthUserContextValueMiddleware` now resolves the auth guard from the config [\#1262 / mfn](https://github.com/rebing/graphql-laravel/pull/1262)